+++ apps/expo/src/app/(tabs)/index.tsx
@@ -1,3 +1,3 @@
-import { HomePage } from "@/src/features/home/home-page";
+import { HomePage } from "@/features/home/home-page";
 
 export default HomePage;
