+++ apps/expo/src/app/(tabs)/settings.tsx
@@ -1,3 +1,3 @@
-import { SettingsPage } from "@/src/features/settings/settings-page";
+import { SettingsPage } from "@/features/settings/settings-page";
 
 export default SettingsPage;
