+++ apps/expo/src/providers/trpc/trpc-provider.tsx
@@ -1,7 +1,6 @@
 import React from "react";
+import { queryClient } from "@/providers/trpc/utils/trpc";
 import { QueryClientProvider } from "@tanstack/react-query";
-
-import { queryClient } from "./utils/trpc";
 
 export function TrpcProvider({ children }: { children: React.ReactNode }) {
   return (
